results/final_count/run-20230925-221829_LLMatic_donnemartin-system-design-oop-whatsapp_0_5-runs_log/whatsapp/1/LLMatic/message.py
import json
import base64


class Message:
	def __init__(self, sender, receiver, content):
		self.sender = sender
		self.receiver = receiver
		self.content = content
		self.read_receipt = False
		self.encrypted = False
		self.image = None

	def send_message(self, messages):
		if self.encrypted:
			self.content = self.decrypt_message()
		messages.append(self)
		return messages

	def receive_message(self, messages):
		for message in messages:
			if message.receiver == self.receiver:
				message.read_receipt = True
		return messages

	def set_read_receipt(self):
		self.read_receipt = True

	def encrypt_message(self):
		self.content = base64.b64encode(self.content.encode()).decode()
		self.encrypted = True

	def decrypt_message(self):
		return base64.b64decode(self.content.encode()).decode()

	def share_image(self, image):
		self.image = image

